pyproject.toml
[tool.poetry]
name = "gpt-api"
version = "0.1.0"
description = ""
authors = ["Akagi-Kazeno <akagikazeno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = {extras = ["async"], version = "^2.3.2"}
openai = "^0.27.7"
requests = "^2.31.0"
opencv-python = "^4.7.0.72"
pillow = "^9.5.0"
sqlalchemy = "^2.0.15"
python-dotenv = "^1.0.0"
flask-cors = "^3.0.10"
numpy = "^1.24.3"
psycopg2-binary = "^2.9.6"
ratelimit = "^2.2.1"
setuptools = "^67.8.0"
aiohttp = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
